Moved Auth service to services/auth
